Fixed Error in chargeCalculationGlobal; Seperated sequenceMap
FIxed an error in chargeCalculationGlobal that improperly coerced a vector into a list.

Seperated sequenceMap by adding a new function that finds the coordinates of residues/positions within the ggplot coordinate plane. This is to reduce complexity of sequenceMap and allows users to output a dataf frame to find coordinates of the map to add additional annotations.