Updating chargeHydropathyPlot
CHP would plot each individual residue in a sequence as a vector of individual amino acids. CHP now takes an icoming vector of individual residues and collapses it into a single amino acid sequence.  Multiple sequences are not supported in this format, but are still accepted as character strings or as path(s) to fasta files